+++ addon/components/flexberry-simpledatetime.js
@@ -4,6 +4,7 @@
 
 import Ember from 'ember';
 import FlexberryBaseComponent from './flexberry-base-component';
+import { translationMacro as t } from 'ember-i18n';
 
 /**
   Wrapper for input[type='date/datetime/datetime-local'] component.
@@ -12,7 +13,7 @@ import FlexberryBaseComponent from './flexberry-base-component';
   @example
     ```handlebars
     {{flexberry-simpledatetime
-      type='datetime-local'
+      type="datetime-local"
       value=model.orderDate
       min=model.orderDateMin
       max=model.orderDateMax
@@ -34,8 +35,11 @@ export default FlexberryBaseComponent.extend({
   _valueAsString: Ember.computed('value', {
     get() {
       let date = this.get('value');
-      let str = this._convertDateToString(date);
-      return str;
+      if (this.get('_supportDateType')) {
+        return this._convertDateToString(date);
+      } else {
+        return date;
+      }
     },
     set(key, value) {
       let date = this._convertStringToDate(value);
@@ -55,8 +59,11 @@ export default FlexberryBaseComponent.extend({
   _minAsString: Ember.computed('min', {
     get() {
       let date = this.get('min');
-      let str = this._convertDateToString(date);
-      return str;
+      if (this.get('_supportDateType')) {
+        return this._convertDateToString(date);
+      } else {
+        return date;
+      }
     },
     set(key, value) {
       let date = this._convertStringToDate(value);
@@ -76,14 +83,25 @@ export default FlexberryBaseComponent.extend({
   _maxAsString: Ember.computed('max', {
     get() {
       let date = this.get('max');
-      let str = this._convertDateToString(date);
-      return str;
+      if (this.get('_supportDateType')) {
+        return this._convertDateToString(date);
+      } else {
+        return date;
+      }
     },
     set(key, value) {
       let date = this._convertStringToDate(value);
       this.set('max', date);
       return value;
     },
+  }),
+
+  _supportDateType: Ember.computed(function() {
+    if (this._checkInput('date') || this._checkInput('datetime') || this._checkInput('datetime-local')) {
+      return true;
+    }
+
+    return false;
   }),
 
   /**
@@ -119,6 +137,91 @@ export default FlexberryBaseComponent.extend({
     @readOnly
   */
   classNames: ['flexberry-simpledatetime'],
+
+  /**
+    Text to be displayed in field, if field has not been filled.
+
+    @property placeholder
+    @type String
+    @default 't('components.flexberry-datepicker.placeholder')'
+  */
+  placeholder: t('components.flexberry-datepicker.placeholder'),
+
+  /**
+    Flatpickr options.
+    For more information see [flatpickr](https://chmln.github.io/flatpickr/)
+  */
+  dateFormat: 'Y-m-dTH:iZ',
+  timeFormat: 'H:i',
+  noCalendar: false,
+  enableTime: true,
+  enableSeconds: false,
+  time_24hr: true,
+  utc: true,
+  altInput: true,
+  altFormat: 'd.m.Y H:i',
+  clickOpens: Ember.computed('readonly', function() {
+    if (this.get('readonly')) {
+      return false;
+    }
+
+    return true;
+  }),
+
+  /**
+    Called when the element of the view has been inserted into the DOM or after the view was re-rendered.
+    For more information see [didInsertElement](http://emberjs.com/api/classes/Ember.Component.html#event_didInsertElement) event of [Ember.Component](http://emberjs.com/api/classes/Ember.Component.html).
+  */
+  didInsertElement() {
+    let _this = this;
+    this._super(...arguments);
+
+    if (!this.get('_supportDateType')) {
+      this._setFlatpickrOptionsWithType();
+      this.$('.flatpickr')[0].flatpickr({
+        dateFormat: this.get('dateFormat'),
+        timeFormat: this.get('timeFormat'),
+        defaultDate: this.get('_valueAsString'),
+        noCalendar: this.get('noCalendar'),
+        enableTime: this.get('enableTime'),
+        enableSeconds: this.get('enableSeconds'),
+        time_24hr: this.get('time_24hr'),
+        utc: this.get('utc'),
+        minDate: this.get('_minAsString'),
+        maxDate: this.get('_maxAsString'),
+        altInput: this.get('altInput'),
+        altFormat: this.get('altFormat'),
+        clickOpens: this.get('clickOpens'),
+
+        // Needs for support IE.
+        onChange: function(dateObj, dateStr, instance) {
+          let date = _this._convertStringToDate(dateStr);
+          _this.set('value', date);
+        }
+      });
+    }
+  },
+
+  /**
+    This method set flatpickr options depending on the type.
+
+    @method _setFlatpickrOptionsWithType
+    @private
+  */
+  _setFlatpickrOptionsWithType() {
+    let type = this.get('type');
+    switch (type) {
+      case 'datetime-local':
+        this.set('utc', true);
+        break;
+      case 'datetime':
+        break;
+      case 'date':
+        this.set('altFormat', 'd.m.Y');
+        this.set('enableTime', false);
+        break;
+    }
+  },
 
   /**
     Convert Date object to appropriate string value for input.
@@ -178,5 +281,19 @@ export default FlexberryBaseComponent.extend({
     }
 
     return new Date(value);
+  },
+
+  /**
+    The method checks if some input type is supported by the browser.
+
+    @method _checkInput
+    @param {String} type Type of input.
+    return {Boolean}
+    @private
+  */
+  _checkInput(type) {
+    let input = document.createElement('input');
+    input.setAttribute('type', type);
+    return input.type === type;
   },
 });