+++ addon/mixins/flexberry-groupedit-route.js
@@ -1,4 +1,5 @@
 import Ember from 'ember';
+import needSaveCurrentAgregator from '../utils/need-save-current-agregator';
 
 /**
   Mixin for {{#crossLink "DS.Route"}}Route{{/crossLink}}
@@ -81,7 +82,6 @@ export default Ember.Mixin.create({
           record = modelToAdd;
         }
 
-        let recordId = record.get('id');
         _this.controller.set('modelNoRollBack', true);
 
         let flexberryDetailInteractionService = _this.get('flexberryDetailInteractionService');
@@ -92,28 +92,33 @@ export default Ember.Mixin.create({
         flexberryDetailInteractionService.pushValue(
           'modelCurrentAgregators', _this.controller.get('modelCurrentAgregators'), _this.controller.get('model'));
 
-        if (recordId) {
-          _this.transitionTo(editFormRoute, record.get('id'))
-          .then(function(newRoute) {
+        if (record.get('isNew')) {
+          let newModelPath = _this.newRoutePath(editFormRoute);
+          _this.transitionTo(newModelPath).then((newRoute) => {
             newRoute.controller.set('readonly', methodOptions.readonly);
           });
         } else {
-          let newModelPath = _this.newRoutePath(editFormRoute);
-          _this.transitionTo(newModelPath)
-          .then(function(newRoute) {
+          _this.transitionTo(editFormRoute, record.get('id')).then((newRoute) => {
             newRoute.controller.set('readonly', methodOptions.readonly);
           });
         }
       };
 
       if (saveBeforeRouteLeave) {
-        this.controller.save().then(() => {
+        this.controller.save(false, true).then(() => {
           goToOtherRouteFunction();
         }).catch((errorData) => {
-          this.rejectError(errorData, this.get('i18n').t('edit-form.save-failed-message'));
+          this.controller.rejectError(errorData, this.get('i18n').t('forms.edit-form.save-failed-message'));
         });
       } else {
         goToOtherRouteFunction();
+      }
+    },
+
+    saveAgregator(agregatorModel) {
+      let agregator = agregatorModel ? agregatorModel : this.get('controller.model');
+      if (needSaveCurrentAgregator.call(this, agregator)) {
+        agregator.save();
       }
     }
   },