+++ addon/routes/list-form.js
@@ -42,12 +42,21 @@ import ReloadListMixin from '../mixins/reload-list-mixin';
   @uses FlexberryObjectlistviewRouteMixin
 */
 export default ProjectedModelFormRoute.extend(
-  PaginatedRouteMixin,
-  SortableRouteMixin,
-  LimitedRouteMixin,
-  ReloadListMixin,
-  FlexberryObjectlistviewRouteMixin,
-  FlexberryObjectlistviewHierarchicalRouteMixin, {
+PaginatedRouteMixin,
+SortableRouteMixin,
+LimitedRouteMixin,
+ReloadListMixin,
+FlexberryObjectlistviewRouteMixin,
+FlexberryObjectlistviewHierarchicalRouteMixin, {
+  /**
+    Link on {{#crossLink FormLoadTimeTrackerService}}{{/crossLink}}.
+
+    @property formLoadTimeTracker
+    @type FormLoadTimeTrackerService
+    @private
+  */
+  formLoadTimeTracker: Ember.inject.service(),
+
   /**
     Current sorting.
 
@@ -66,6 +75,7 @@ export default ProjectedModelFormRoute.extend(
     @param {Object} transition
   */
   model: function(params, transition) {
+    this.get('formLoadTimeTracker').set('startLoadTime', performance.now());
     let modelName = this.get('modelName');
     let webPage = transition.targetName;
     let projectionName = this.get('modelProjection');
@@ -123,25 +133,35 @@ export default ProjectedModelFormRoute.extend(
           page: params.page,
           sorting: this.sorting,
           filter: params.filter,
+          filterCondition: this.get('controller.filterCondition'),
           filters: filtersPredicate,
           predicate: limitPredicate,
           hierarchicalAttribute: hierarchicalAttribute,
         };
+
+        this.onModelLoadingStarted(queryParameters);
 
         // Find by query is always fetching.
         // TODO: support getting from cache with "store.all->filterByProjection".
         // TODO: move includeSorting to setupController mixins?
         return this.reloadList(queryParameters);
       }).then((records) => {
+        this.get('formLoadTimeTracker').set('endLoadTime', performance.now());
+        this.onModelLoadingFulfilled(records);
         this.includeSorting(records, this.sorting);
         this.get('controller').set('model', records);
         return records;
+      }).catch((errorData) => {
+        this.onModelLoadingRejected(errorData);
+      }).finally((data) => {
+        this.onModelLoadingAlways(data);
       });
 
     if (this.get('controller') === undefined) {
       return { isLoading: true };
     }
 
+    // TODO: Check controller loaded model loading parameters and return it without reloading if there is same backend query was executed.
     let model = this.get('controller.model');
 
     if (model !== null) {
@@ -149,6 +169,73 @@ export default ProjectedModelFormRoute.extend(
     } else {
       return { isLoading: true };
     }
+  },
+
+  /**
+    This method will be invoked before model loading operation will be called.
+    Override this method to add some custom logic on model loading operation start.
+
+    @example
+      ```javascript
+      onModelLoadingStarted(queryParameters) {
+        alert('Model loading operation started!');
+      }
+      ```
+    @method onModelLoadingStarted.
+    @param {Object} queryParameters Query parameters used for model loading operation.
+  */
+  onModelLoadingStarted(queryParameters) {
+  },
+
+  /**
+    This method will be invoked when model loading operation successfully completed.
+    Override this method to add some custom logic on model loading operation success.
+
+    @example
+      ```javascript
+      onModelLoadingFulfilled() {
+        alert('Model loading operation succeed!');
+      }
+      ```
+    @method onModelLoadingFulfilled.
+    @param {Object} model Loaded model data.
+  */
+  onModelLoadingFulfilled(model) {
+  },
+
+  /**
+    This method will be invoked when model loading operation completed, but failed.
+    Override this method to add some custom logic on model loading operation fail.
+
+    @example
+      ```javascript
+      onModelLoadingRejected() {
+        alert('Model loading operation failed!');
+      }
+      ```
+    @method onModelLoadingRejected.
+    @param {Object} errorData Data about model loading operation fail.
+  */
+  onModelLoadingRejected(errorData) {
+    // TODO: Provide information about error to user.
+  },
+
+  /**
+    This method will be invoked always when model loading operation completed,
+    regardless of model loading promise's state (was it fulfilled or rejected).
+    Override this method to add some custom logic on model loading operation completion.
+
+    @example
+      ```js
+      onModelLoadingAlways(data) {
+        alert('Model loading operation completed!');
+      }
+      ```
+
+    @method onModelLoadingAlways.
+    @param {Object} data Data about completed model loading operation.
+  */
+  onModelLoadingAlways(data) {
   },
 
   /**
@@ -161,6 +248,7 @@ export default ProjectedModelFormRoute.extend(
   */
   setupController: function(controller, model) {
     this._super(...arguments);
+    this.get('formLoadTimeTracker').set('startRenderTime', performance.now());
 
     // Define 'modelProjection' for controller instance.
     // TODO: remove that when list-form controller will be moved to this route.
@@ -168,5 +256,5 @@ export default ProjectedModelFormRoute.extend(
     let proj = modelClass.projections.get(this.get('modelProjection'));
     controller.set('userSettings', this.userSettings);
     controller.set('modelProjection', proj);
-  },
+  }
 });