+++ blueprints/ember-flexberry/index.js
@@ -11,6 +11,7 @@ module.exports = {
 
     return this.addBowerPackagesToProject([
       { name: 'semantic-ui-daterangepicker', target: '5d46ed2e6e5a0bf398bb6a5df82e06036dfc46be' },
+      { name: 'flatpickr-calendar', source: 'git://github.com/chmln/flatpickr.git', target: '4ca9590caa70bc0232cf0a3455cfac1be9c65c2a' },
       { name: 'blueimp-file-upload', target: '9.11.2' },
       { name: 'devicejs', target: '0.2.7' }
     ]).then(function() {
@@ -19,6 +20,7 @@ module.exports = {
           { name: 'semantic-ui-ember', target: '0.9.3' },
           { name: 'ember-moment', target: '6.0.0' },
           { name: 'ember-link-action', target: '0.0.35' },
+          { name: 'ember-href-to', target: '1.9.0' },
           { name: 'ember-cli-sass', target: '5.2.0' },
           { name: 'broccoli-jscs', target: '1.2.2' },
           { name: 'ember-browserify', target: '1.1.9' }